pyproject.toml
[tool.poetry]
name = "oer-scraper"
version = "0.1.0"
description = "Scrape German public television programs."
authors = ["Philipp Kessling <p.kessling@leibniz-hbi.de>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "*"
pandas = "*"
click = "*"
tqdm = "*"
pyyaml = "*"
pydantic = "*"
loguru = "*"

[tool.poetry.scripts]
"oerscraper" = "oer_scraper.cli:cli"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
